requirements.txt
Django==4.2.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
